requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.21
pymssql==2.2.8
python-dotenv==1.0.0
Werkzeug==2.3.7 